+++ ExilePartyClient/main.ts
@@ -1,10 +1,22 @@
-import { app, BrowserWindow, screen } from 'electron';
+import { app, BrowserWindow, dialog, screen } from 'electron';
 import * as path from 'path';
 import * as url from 'url';
 
+const log = require('electron-log');
+const { autoUpdater } = require('electron-updater');
 let win, serve;
 const args = process.argv.slice(1);
 serve = args.some(val => val === '--serve');
+
+autoUpdater.logger = log;
+autoUpdater.logger.transports.file.level = 'info';
+log.info('App starting...');
+
+
+function sendStatusToWindow(text) {
+  log.info(text);
+  win.webContents.send('message', text);
+}
 
 function createWindow() {
 
@@ -17,9 +29,11 @@ function createWindow() {
     y: 100,
     width: 1341,
     height: 987,
+    minHeight: 987,
+    minWidth: 1341,
     webPreferences: { webSecurity: false },
     frame: false,
-    resizable: false,
+    resizable: true,
     icon: path.join(__dirname, 'dist/assets/img/app-icon.png'),
   });
 
@@ -46,11 +60,39 @@ function createWindow() {
 }
 
 try {
+  autoUpdater.on('checking-for-update', () => {
+    sendStatusToWindow('Checking for update...');
+  });
+  autoUpdater.on('update-available', (info) => {
+    dialog.showMessageBox({ title: 'Update available!', message: 'Press OK to download and apply the update.' });
+    sendStatusToWindow('Update available.');
+  });
+  autoUpdater.on('update-not-available', (info) => {
+    sendStatusToWindow('Update not available.');
+  });
+  autoUpdater.on('error', (err) => {
+    sendStatusToWindow('Error in auto-updater. ' + err);
+  });
+  autoUpdater.on('download-progress', (progressObj) => {
+    let log_message = 'Download speed: ' + progressObj.bytesPerSecond;
+    log_message = log_message + ' - Downloaded ' + progressObj.percent + '%';
+    log_message = log_message + ' (' + progressObj.transferred + '/' + progressObj.total + ')';
+    sendStatusToWindow(log_message);
+  });
+
+
+  autoUpdater.on('update-downloaded', (info) => {
+    sendStatusToWindow('Update downloaded');
+    autoUpdater.quitAndInstall();
+  });
 
   // This method will be called when Electron has finished
   // initialization and is ready to create browser windows.
   // Some APIs can only be used after this event occurs.
-  app.on('ready', createWindow);
+  app.on('ready', () => {
+    createWindow();
+    autoUpdater.checkForUpdates();
+  });
 
   // Quit when all windows are closed.
   app.on('window-all-closed', () => {
@@ -73,3 +115,5 @@ try {
   // Catch Error
   // throw e;
 }
+
+