+++ ExilePartyClient/src/app/authorize/authorize.module.ts
@@ -1,12 +1,14 @@
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule, MatFormFieldModule, MatIconModule, MatInputModule, MatCheckboxModule } from '@angular/material';
+import { MatButtonModule, MatFormFieldModule, MatIconModule, MatInputModule } from '@angular/material';
 import { MatListModule } from '@angular/material/list';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { RouterModule } from '@angular/router';
 
+import { PartyService } from '../shared/providers/party.service';
 import { SharedModule } from '../shared/shared.module';
 import { AuthorizeComponent } from './authorize.component';
+import { DashboardModule } from './dashboard/dashboard.module';
 import { PartyModule } from './party/party.module';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { PartyService } from '../shared/providers/party.service';