+++ ExilePartyClient/src/app/login/login.module.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { MatButtonModule, MatInputModule, MatOptionModule, MatSelectModule, MatIconModule } from '@angular/material';
+import { MatButtonModule, MatInputModule, MatOptionModule, MatSelectModule, MatIconModule,
+  MatStepperModule, MatProgressBarModule } from '@angular/material';
 import { MatFormFieldModule } from '@angular/material/form-field';
 
 import { LoginComponent } from './login.component';
@@ -14,10 +15,12 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
     ReactiveFormsModule,
     MatInputModule,
     MatIconModule,
+    MatStepperModule,
     MatButtonModule,
     MatOptionModule,
     MatSelectModule,
-    MatFormFieldModule
+    MatFormFieldModule,
+    MatProgressBarModule
   ],
   declarations: [LoginComponent]
 })